+++ examples/autosnag.mjs
@@ -1,0 +1,13 @@
+import Bot from 'ttapi';
+
+var AUTH   = process.env.TTAPI_AUTH || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var USERID = process.env.TTAPI_USER || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var ROOMID = process.env.TTAPI_ROOM || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+
+var bot = new Bot(AUTH, USERID, ROOMID);
+
+bot.on('newsong', function (data) {
+  bot.snag();
+  bot.playlistAdd(data.room.metadata.current_song._id);
+  bot.becomeFan(data.room.metadata.current_dj);
+});
