+++ examples/dj_bot.mjs
@@ -1,7 +1,8 @@
-var Bot = require('ttapi');
-var AUTH = 'xxxxxxxxxxxxxxxxxxxxxxxx';
-var USERID = 'xxxxxxxxxxxxxxxxxxxxxxxx';
-var ROOMID = 'xxxxxxxxxxxxxxxxxxxxxxxx';
+import Bot from 'ttapi';
+
+var AUTH   = process.env.TTAPI_AUTH || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var USERID = process.env.TTAPI_USER || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var ROOMID = process.env.TTAPI_ROOM || 'xxxxxxxxxxxxxxxxxxxxxxxx';
 
 var bot = new Bot(AUTH, USERID, ROOMID);
