+++ examples/room_list.mjs
@@ -1,0 +1,16 @@
+import Bot from 'ttapi';
+
+var AUTH   = process.env.TTAPI_AUTH || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var USERID = process.env.TTAPI_USER || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var ROOMID = process.env.TTAPI_ROOM || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+
+var bot = new Bot(AUTH, USERID);
+
+var theUsersList = {};
+
+bot.once('ready', function() {
+  console.log("ready");
+  bot.directoryRooms(function(data) {
+    console.log(data);
+  })
+});
