+++ examples/switch.mjs
@@ -2,10 +2,11 @@
  * On/Off bot switch with basic variables in nodejs.
  */
 
-var Bot    = require('ttapi')
-  , AUTH   = 'xxxxxxxxxxxxxxxxxxxxxxxx'
-  , USERID = 'xxxxxxxxxxxxxxxxxxxxxxxx'
-  , ROOMID = 'xxxxxxxxxxxxxxxxxxxxxxxx';
+import Bot from 'ttapi';
+
+var AUTH   = process.env.TTAPI_AUTH || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var USERID = process.env.TTAPI_USER || 'xxxxxxxxxxxxxxxxxxxxxxxx';
+var ROOMID = process.env.TTAPI_ROOM || 'xxxxxxxxxxxxxxxxxxxxxxxx';
 
 var bot = new Bot(AUTH, USERID, ROOMID);
